CHG: Add test cases for setting ANX.
- Test ScenarioError raising when no ANX given at all.
- Test getting ANX from orbit prediction file if not in scenario.
- Test getting ANX from scenario.
- Test scenario ANX precedence over orbit prediction file.
